Add warning for git actions
There are now multiple commands which make use of git commands
to create output diffs between brew branches. This is potentially
harmful if the user is running a brew command with another
branch or has changes already. We try to check for those things
but it's not 100% safe. So we should warn users about that behavior.