Let brew install gems automatically in CI
Bo made some really nice changes to how gems are installed in
brew so I think it will do a better job of choosing to install
only the gems necessary for the command that's getting called.

This means we can remove manual gem installation and trust
that brew will just do the right thing.