add brew grep command
This is an exploration of statically parsing the cask/formula files
using grep + regexes to if we it is fast enough to allow us to parse
that information on the fly each time and see if it's accurate enough
to make the old command obsolete. The old technique required caching
descriptions which is extra complexity and is a security risk since
it means that all package files are evaluated.

As is this is super hacky but it does show that this is a viable
solution to the problem.