Fix cask token error in `brew api-readall-test`
```
$ brew api-readall-test
Generating formulae API ... and mocking formula API loader.
Read core formulae and saw no failures!

Generating cask API ...Warning: Cask font-rounded-mplus was renamed to font-rounded-m+.
 and mocking cask API loader.
Read core casks and saw 1 failure!

Readall Failures: Casks
-----------------------
>> font-rounded-mplus : key not found: "font-rounded-mplus"
/usr/local/Homebrew/Library/Homebrew/cask/cask_loader.rb:318:in `fetch'
```

The problem was that `font-rounded-mplus` got renamed to `font-rounded-m+`
and we were getting the cask tokens from the file names instead of the cask
definitions directly. I'm not entirely sure why the cask token name from
the core cask tap are wrong here but changing the order and basing it on
the cask that was loaded seems to work.