Fix service-diff command
The problem here was that for some reason on CI the hashicorp tap
is installed and that means that I was running this command for that
tap too. This didn't really make any sense because we only really
need to run this command for the core tap; nothing else gets serialized
to the API.

I removed the --tap= command and now it just runs on the core tap by
default which should help us avoid this problem.

I tested it both with all formulae in the core tap and a single formula.